requirements.txt
Django >=2.1.3,<=2.2.0
requests >= 2.18.4,<=2.27.1
wheel 
git+https://github.com/de-laz/openhim-mediator-utils-py
djangorestframework == 3.10.3
django-import-export
APScheduler==3.8.1
python-dotenv<=0.17.1